Add files via upload
This is the original demo for P.H.A.R.M.A.C.E.U.T.I.C.A.L. The program allows the user to gain entry to the program via password. The user then gets to simulate the responsibilities of a pharmacy technician by viewing electronic prescriptions, accessing patient records, as well as ordering medications from a wholesaler which results in the creation of an invoice. If the user accesses the program as a pharmacist, the user is permitted to add pharmacy technicians into the registrar as well as ordering controlled substances.