Add files via upload
P.H.A.R.M.A.C.E.U.T.I.C.A.L. Patch 1 expands upon the base premise of the demo, improving the simulation experience with additional features for both pharmacist and technician roles, improvised role-based access controls, as well as data retention of patient information, a staff registrar, and logging records of all employees that access the program. Data persistence has been implemented to keep changes made to any database after logging out or closing the program. The code in Patch 1 is also much more efficient with the creation of Pharmacist and Technician classes which feature methods and properties to make the code easier to understand,